renamed Nav to nav for consistency sake
